refactor: #675 chat_question_generator を生成/フォールバック境界で分割する
戦略質問・AI生成・フォールバック・事前定義質問をファイル分割し、
単一ファイルの巨大化を解消する。公開 API・質問契約は不変。
